% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{curr_state}
\alias{curr_state}
\title{Current state}
\usage{
curr_state()
}
\value{
"curr_state"
}
\description{
A function used inside event_mapping(probs = ) placeholder for returning the current state
}
\examples{
curr_state()
}
