% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{decisions}
\alias{decisions}
\title{Add decisions to a gmod}
\usage{
decisions(...)
}
\arguments{
\item{...}{decision names}
}
\value{
a gmod layer with decision names
}
\description{
Add decisions to a gmod
}
\examples{
decisions("A", "B", "C")
}
