% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{discounts}
\alias{discounts}
\title{Add discounts to a gmod Markov object}
\usage{
discounts(payoffs, discounts)
}
\arguments{
\item{...}{decision names}
}
\value{
a gmod layer with decision names
}
\description{
Add discounts to a gmod Markov object
}
\examples{
discounts(payoffs = c("cost", "effectiveness"), discounts = c(0.5, 0.5))
discounts(payoffs = c("cost", "effectiveness"), discounts = c(0.15, 0.15))
}
