% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{gmod}
\alias{gmod}
\title{Create a new gmod}
\usage{
gmod(name, model_function_name = NULL)
}
\arguments{
\item{model_type}{}
}
\value{
a new gmod object
}
\description{
Create a new gmod
}
\examples{
gmod(model_type = "Markov", n_cycles = 10)

}
