% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_build.R
\name{gmod_build.gmod_decision}
\alias{gmod_build.gmod_decision}
\title{Build a decision tree structure}
\usage{
\method{gmod_build}{gmod_decision}(gmod_obj)
}
\arguments{
\item{gmod_obj}{}
}
\value{
gmod structure containing formulae strings
}
\description{
takes in a gmod model syntax and returns a gmod formula structure
}
