% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_build.R
\name{gmod_build.gmod_markov}
\alias{gmod_build.gmod_markov}
\title{Build a Markov model structure}
\usage{
\method{gmod_build}{gmod_markov}(gmod_obj, params = NULL, simplify = FALSE)
}
\arguments{
\item{gmod_obj}{}
}
\value{
gmod structure containing formulae strings
}
\description{
takes in a gmod formula structure and returns a gmod numerical formula structure
}
\examples{
print("see vignettes(package = 'gmod')") 
}
