% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_build.R
\name{gmod_build}
\alias{gmod_build}
\title{Builds a Markov model}
\usage{
gmod_build(x, ...)
}
\arguments{
\item{gmod_obj}{}
}
\value{
a gmod object containing the equations for the Markov model
}
\description{
Builds a Markov model
}
\examples{
gmod_build(mygmod)
}
