% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_evaluate.R
\name{gmod_evaluate.gmod_decision}
\alias{gmod_evaluate.gmod_decision}
\title{Runs the decision tree}
\usage{
\method{gmod_evaluate}{gmod_decision}(model_struc, params = NULL)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical gmod decision object from the gmod_parse() function}
}
\value{
gmod object
}
\description{
runs the decision model and returns the expected payoff values and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='gmod')")
}
