% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_evaluate.R
\name{gmod_evaluate}
\alias{gmod_evaluate}
\title{Runs a decision tree or Markov model}
\usage{
gmod_evaluate(x, ...)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical gmod object from the gmod_parse() function}
}
\value{
gmod object
}
\description{
runs the decision tree or markov model and returns the outcomes either traces and summary outcomes
}
\examples{
gmod_evaluate(numerical_model_structure)
}
