% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_evaluate_functions.R
\name{gmod_evaluate_functions}
\alias{gmod_evaluate_functions}
\title{Title}
\usage{
gmod_evaluate_functions(
  gmod_obj,
  n_cycles = NULL,
  fun_names = NULL,
  excel_file_name = NULL
)
}
\arguments{
\item{gmod_obj}{}

\item{fun_names}{}

\item{excel_file_name}{}
}
\value{
TRUE
}
\description{
Title
}
\examples{
TRUE

}
