% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_expand_functions.R
\name{gmod_expand_functions}
\alias{gmod_expand_functions}
\title{Title}
\usage{
gmod_expand_functions(gmod_obj, fun_names, excel_file_name = NULL)
}
\arguments{
\item{gmod_obj}{}

\item{fun_names}{}

\item{excel_file_name}{}
}
\description{
Title
}
