% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_gen_model_function.R
\name{gmod_gen_model_function.gmod_decision}
\alias{gmod_gen_model_function.gmod_decision}
\title{Title}
\usage{
\method{gmod_gen_model_function}{gmod_decision}(
  model_struc,
  model_function_name = "my_decision_model",
  print_model_function = FALSE,
  return_model_structure = TRUE
)
}
\arguments{
\item{model_struc}{}

\item{model_function_name}{}

\item{print_model_function}{}
}
\value{
model strucuture and generates the model function
}
\description{
Title
}
\examples{
print("vignette(package = 'gmod')")
}
