% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_gen_model_function.R
\name{gmod_gen_model_function}
\alias{gmod_gen_model_function}
\title{Runs a decision tree or Markov model}
\usage{
gmod_gen_model_function(x, ...)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical gmod object from the gmod_parse() function}
}
\value{
model_struc and generates model function code
}
\description{
runs the decision tree or markov model and returns the outcomes either traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='gmod')")
}
