% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_parse.R
\name{gmod_parse.gmod_markov}
\alias{gmod_parse.gmod_markov}
\title{Parses a gmod Markov object}
\usage{
\method{gmod_parse}{gmod_markov}(model_struc, params = NULL)
}
\arguments{
\item{model_struc}{a gmod representing the model structure}

\item{params}{a parameter list containing the model parameters}
}
\value{
model_num_struc
}
\description{
given gmod equations and a list of parameters it evaluates the model equations and returns the numerical values
}
\examples{
print("See Vignettes")
}
