% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_parse.R
\name{gmod_parse}
\alias{gmod_parse}
\title{Parses a gmod model formula structure}
\usage{
gmod_parse(x, ...)
}
\arguments{
\item{gmod_obj}{a gmod object containing the equations for the Markov model}

\item{params}{a list containing the model parameters}
}
\value{
a nuemerical gmod object structure
}
\description{
parses the string formulae and computes the numerical outcomes
}
\examples{
gmod_parse(mygmod, params = list(param1 = 0.5, param2 = 0.6))
}
