% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Title}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
TRUE if x is not in y, otherwise false
}
\description{
Title
}
\examples{
1 \%not_in\% 
}
