% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{initial_probs}
\alias{initial_probs}
\title{Add initial probabilities to a Markov gmod object}
\usage{
initial_probs(states, probs)
}
\arguments{
\item{states}{state names for which we assign initial probabilities}

\item{probs}{the initial probabilities corresponding to the states}
}
\value{
a gmod layer with initial probabilities
}
\description{
Add initial probabilities to a Markov gmod object
}
\examples{
initial_probs(states = c("Healthy", "Sick", "Dead"), probs = c(1,0,0))
initial_probs(states = c("Healthy", "Sick", "Dead"), probs = c(0.5,0.3,Inf))
initial_probs(states = "Healthy", probs = 1)
}
