% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{outcomes}
\alias{outcomes}
\title{Add outcomes from a decision tree to a gmod}
\usage{
outcomes(...)
}
\arguments{
\item{...}{Decision outcome names}
}
\value{
a gmod layer with Decision Outcome names
}
\description{
Add outcomes from a decision tree to a gmod
}
\examples{
outcomes("Alive", "Dead")
}
