% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{payoffs}
\alias{payoffs}
\title{Add final_outcomes from a decision tree to a gmod}
\usage{
payoffs(...)
}
\arguments{
\item{...}{a named list containing the payoffs and the associated payoff functions}
}
\value{
a gmod layer with Decision final_outcome names

a gmod layer with payoffs
}
\description{
Add final_outcomes from a decision tree to a gmod
}
\examples{
final_outcomes("Alive", "Dead")
Add payoffs to a gmod

payoffs(cost = cost_function(state), effectiveness = effective_function(state))
}
