% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{+.gmod_class}
\alias{+.gmod_class}
\title{Define a method for the \code{+} operator for \code{gmod} objects}
\usage{
\method{+}{gmod_class}(gmod_obj, layer)
}
\arguments{
\item{gmod_obj}{}

\item{layer}{}
}
\value{
gmod_obj
}
\description{
Define a method for the \code{+} operator for \code{gmod} objects
}
