% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{prev_event}
\alias{prev_event}
\title{Value of a previous event}
\usage{
prev_event(...)
}
\value{
"curr_state"
}
\description{
A function used inside event_mapping(probs = ) placeholder for returning the value of the previous event
}
\examples{
curr_state() 
}
