% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{print.gmod_class}
\alias{print.gmod_class}
\title{Define a print method for \code{gmod_class} objects}
\usage{
\method{print}{gmod_class}(gmod_obj)
}
\arguments{
\item{gmod_obj}{The gmod object}
}
\description{
Define a print method for \code{gmod_class} objects
}
