% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{prob2rate}
\alias{prob2rate}
\title{Title}
\usage{
prob2rate(prob)
}
\arguments{
\item{prob}{}
}
\value{
rate
}
\description{
converts prob to rate using rate = -log(1-prob)
}
\examples{
prob2rate(prob = 0.5)
}
