% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{prob_left}
\alias{prob_left}
\title{Complementary probability}
\usage{
prob_left()
}
\value{
"curr_state"
}
\description{
A function used inside event_mapping(probs = ) placeholder for returning Inf. which will be used internally to return the complementary probabilities
}
\examples{
curr_state() 
}
