% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{rate2prob}
\alias{rate2prob}
\title{Title}
\usage{
rate2prob(rate)
}
\arguments{
\item{rate}{}
}
\value{
probability
}
\description{
converts rate to probability using prob = 1-exp(-rate)
}
\examples{
rate2prob(rate = 0.3)
}
