% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{tunnel2state}
\alias{tunnel2state}
\title{Title}
\usage{
tunnel2state(tunnel_state)
}
\arguments{
\item{tunnel_state}{the name of the tunnel state}
}
\value{
list consisting of the name of the state and it's cycle_in_state value
}
\description{
decomposes an expanded state into the state name and cycle_in_state. If the state is not a tunnel, it will return cycle_in_state=1
}
\examples{
tunnel2state("sick_tnl5")
}
