% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{tunnels}
\alias{tunnels}
\title{Title}
\usage{
tunnels(states, lengths = NULL)
}
\arguments{
\item{states}{names of the states to expand}

\item{lengths}{the length of each tunnel state}
}
\value{
a gmod layer with tunnels
}
\description{
Title
}
\examples{
tunnels(states = c("S1", "S2"), lengths = c(3, 5))
}
