% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{event_mapping}
\alias{event_mapping}
\title{Add event mapping}
\usage{
event_mapping(event, values, results, probs, payoffs = NULL)
}
\arguments{
\item{event}{}

\item{values}{}

\item{results}{}

\item{probs}{}

\item{payoffs}{}
}
\value{
gmod layer
}
\description{
Add event mapping
}
\examples{
event_mapping(event = "event_progress", 
values = c(TRUE,FALSE), 
results = c("Severe",curr_state()), 
probs = c(p_progress_function(state), Inf), 
payoffs = list(cost = c(cProgress, cNotProgress)))
}
