% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{gmod}
\alias{gmod}
\title{Create a new gmod}
\usage{
gmod()
}
\arguments{
\item{model_type}{}

\item{n_cycles}{= 50}
}
\value{
a new gmod object
}
\description{
Create a new gmod
}
\examples{
gmod(model_type = "Markov", n_cycles = 40)

}
