% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_evaluate.R
\name{gmod_evaluate.gmod_markov}
\alias{gmod_evaluate.gmod_markov}
\title{Runs the markov model}
\usage{
\method{gmod_evaluate}{gmod_markov}(
  model_struc,
  model_function_name = "my_markov_model",
  print_model_function = FALSE
)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical gmod object from the gmod_parse() function}
}
\description{
runs the markov model and returns the traces and summary outcomes
}
\examples{
gmod_evaluate(numerical_model_structure)
}
