% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_gen_model_function.R
\name{gmod_gen_model_function.gmod_decision}
\alias{gmod_gen_model_function.gmod_decision}
\title{Title}
\usage{
\method{gmod_gen_model_function}{gmod_decision}(
  gmod_obj,
  model_function_name = "my_decision_model",
  print_model_function = FALSE,
  return_model_structure = TRUE
)
}
\arguments{
\item{gmod_obj}{}

\item{model_function_name}{}

\item{print_model_function}{}
}
\value{
model_struc and generates model function code
}
\description{
Title
}
\examples{
print("see vignettes: vignettes(package='gmod')")
}
