% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmod_gen_model_function.R
\name{gmod_gen_model_function.gmod_markov}
\alias{gmod_gen_model_function.gmod_markov}
\title{Runs the markov model}
\usage{
\method{gmod_gen_model_function}{gmod_markov}(
  mygmod,
  model_function_name = "my_markov_model",
  print_model_function = FALSE,
  sparse = FALSE,
  return_model_structure = TRUE
)
}
\arguments{
\item{model_struc}{a matrix containing the numerical gmod object from the gmod_parse() function}
}
\value{
model strucuture and generates the model function
}
\description{
runs the markov model and returns the traces and summary outcomes
}
\examples{
gmod_evaluate(numerical_model_structure)
}
