% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmod_build.R
\name{final_outcomes}
\alias{final_outcomes}
\title{Add final_outcomes from a decision tree to a gmod}
\usage{
final_outcomes(...)
}
\arguments{
\item{...}{Decision final_outcome names}
}
\value{
a gmod layer with Decision final_outcome names
}
\description{
Add final_outcomes from a decision tree to a gmod
}
\examples{
final_outcomes("Alive", "Dead")
}
